+++ cmd/buildah/config.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/containers/buildah"
+	"github.com/containers/buildah/define"
 	"github.com/containers/buildah/docker"
 	buildahcli "github.com/containers/buildah/pkg/cli"
 	"github.com/mattn/go-shellwords"
@@ -294,7 +295,7 @@ func updateConfig(builder *buildah.Builder, c *cobra.Command, iopts configResult
 	}
 	if c.Flag("hostname").Changed {
 		name := iopts.hostname
-		if name != "" && builder.Format == buildah.OCIv1ImageManifest {
+		if name != "" && builder.Format == define.OCIv1ImageManifest {
 			logrus.Warnf("HOSTNAME is not supported for OCI V1 image format, hostname %s will be ignored. Must use `docker` format", name)
 		}
 		builder.SetHostname(name)
