+++ go.mod
@@ -4,8 +4,8 @@ go 1.12
 
 require (
 	github.com/containernetworking/cni v0.8.1
-	github.com/containers/common v0.35.0
-	github.com/containers/image/v5 v5.10.3
+	github.com/containers/common v0.35.2
+	github.com/containers/image/v5 v5.10.5
 	github.com/containers/ocicrypt v1.1.0
 	github.com/containers/storage v1.27.0
 	github.com/docker/distribution v2.7.1+incompatible
