+++ util/types.go
@@ -1,35 +1,24 @@
 package util
 
+import (
+	"github.com/containers/buildah/define"
+)
+
 const (
-	// Deprecated: Default runtime should come from containers.conf
-	DefaultRuntime = "runc"
+	// DefaultRuntime if containers.conf fails.
+	DefaultRuntime = define.DefaultRuntime
 	// DefaultCNIPluginPath is the default location of CNI plugin helpers.
-	DefaultCNIPluginPath = "/usr/libexec/cni:/opt/cni/bin"
+	DefaultCNIPluginPath = define.DefaultCNIPluginPath
 	// DefaultCNIConfigDir is the default location of CNI configuration files.
-	DefaultCNIConfigDir = "/etc/cni/net.d"
+	DefaultCNIConfigDir = define.DefaultCNIConfigDir
 )
 
 var (
 	// DefaultCapabilities is the list of capabilities which we grant by
 	// default to containers which are running under UID 0.
-	DefaultCapabilities = []string{
-		"CAP_AUDIT_WRITE",
-		"CAP_CHOWN",
-		"CAP_DAC_OVERRIDE",
-		"CAP_FOWNER",
-		"CAP_FSETID",
-		"CAP_KILL",
-		"CAP_MKNOD",
-		"CAP_NET_BIND_SERVICE",
-		"CAP_SETFCAP",
-		"CAP_SETGID",
-		"CAP_SETPCAP",
-		"CAP_SETUID",
-		"CAP_SYS_CHROOT",
-	}
+	DefaultCapabilities = define.DefaultCapabilities
+
 	// DefaultNetworkSysctl is the list of Kernel parameters which we
 	// grant by default to containers which are running under UID 0.
-	DefaultNetworkSysctl = map[string]string{
-		"net.ipv4.ping_group_range": "0 0",
-	}
+	DefaultNetworkSysctl = define.DefaultNetworkSysctl
 )
